# internal/web/olegalentities/api.gen.go
// Package olegalentities provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package olegalentities

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
}

// LegalEntityCreateRequest defines model for LegalEntityCreateRequest.
type LegalEntityCreateRequest struct {
	Name string `json:"name"`
}

// LegalEntityDTO defines model for LegalEntityDTO.
type LegalEntityDTO struct {
	CreatedAt time.Time          `json:"created_at"`
	DeletedAt *time.Time         `json:"deleted_at"`
	Name      string             `json:"name"`
	UpdatedAt *time.Time         `json:"updated_at,omitempty"`
	Uuid      openapi_types.UUID `json:"uuid"`
}

// LegalEntityUpdateRequest defines model for LegalEntityUpdateRequest.
type LegalEntityUpdateRequest struct {
	Name string `json:"name"`
}

// PostLegalEntitiesResponse defines model for PostLegalEntitiesResponse.
type PostLegalEntitiesResponse struct {
	Uuid openapi_types.UUID `json:"uuid"`
}

// EntityUUID defines model for entityUUID.
type EntityUUID = openapi_types.UUID

// Uuid defines model for uuid.
type Uuid = openapi_types.UUID

// PostLegalEntitiesJSONRequestBody defines body for PostLegalEntities for application/json ContentType.
type PostLegalEntitiesJSONRequestBody = LegalEntityCreateRequest

// PutLegalEntitiesUuidJSONRequestBody defines body for PutLegalEntitiesUuid for application/json ContentType.
type PutLegalEntitiesUuidJSONRequestBody = LegalEntityUpdateRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Получить все юридические лица
	// (GET /legal-entities)
	GetLegalEntities(ctx echo.Context) error
	// Создать юридическое лицо
	// (POST /legal-entities)
	PostLegalEntities(ctx echo.Context) error
	// Удалить юридическое лицо
	// (DELETE /legal-entities/{uuid})
	DeleteLegalEntitiesUuid(ctx echo.Context, uuid openapi_types.UUID) error
	// Обновить юридическое лицо
	// (PUT /legal-entities/{uuid})
	PutLegalEntitiesUuid(ctx echo.Context, uuid openapi_types.UUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetLegalEntities converts echo context to params.
func (w *ServerInterfaceWrapper) GetLegalEntities(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetLegalEntities(ctx)
	return err
}

// PostLegalEntities converts echo context to params.
func (w *ServerInterfaceWrapper) PostLegalEntities(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostLegalEntities(ctx)
	return err
}

// DeleteLegalEntitiesUuid converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteLegalEntitiesUuid(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "uuid" -------------
	var uuid openapi_types.UUID

	err = runtime.BindStyledParameterWithLocation("simple", false, "uuid", runtime.ParamLocationPath, ctx.Param("uuid"), &uuid)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter uuid: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteLegalEntitiesUuid(ctx, uuid)
	return err
}

// PutLegalEntitiesUuid converts echo context to params.
func (w *ServerInterfaceWrapper) PutLegalEntitiesUuid(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "uuid" -------------
	var uuid openapi_types.UUID

	err = runtime.BindStyledParameterWithLocation("simple", false, "uuid", runtime.ParamLocationPath, ctx.Param("uuid"), &uuid)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter uuid: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutLegalEntitiesUuid(ctx, uuid)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/legal-entities", wrapper.GetLegalEntities)
	router.POST(baseURL+"/legal-entities", wrapper.PostLegalEntities)
	router.DELETE(baseURL+"/legal-entities/:uuid", wrapper.DeleteLegalEntitiesUuid)
	router.PUT(baseURL+"/legal-entities/:uuid", wrapper.PutLegalEntitiesUuid)

}

type GetLegalEntitiesRequestObject struct {
}

type GetLegalEntitiesResponseObject interface {
	VisitGetLegalEntitiesResponse(w http.ResponseWriter) error
}

type GetLegalEntities200JSONResponse []LegalEntityDTO

func (response GetLegalEntities200JSONResponse) VisitGetLegalEntitiesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetLegalEntities401JSONResponse ErrorResponse

func (response GetLegalEntities401JSONResponse) VisitGetLegalEntitiesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type PostLegalEntitiesRequestObject struct {
	Body *PostLegalEntitiesJSONRequestBody
}

type PostLegalEntitiesResponseObject interface {
	VisitPostLegalEntitiesResponse(w http.ResponseWriter) error
}

type PostLegalEntities201JSONResponse PostLegalEntitiesResponse

func (response PostLegalEntities201JSONResponse) VisitPostLegalEntitiesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type PostLegalEntities401JSONResponse ErrorResponse

func (response PostLegalEntities401JSONResponse) VisitPostLegalEntitiesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type DeleteLegalEntitiesUuidRequestObject struct {
	Uuid openapi_types.UUID `json:"uuid"`
}

type DeleteLegalEntitiesUuidResponseObject interface {
	VisitDeleteLegalEntitiesUuidResponse(w http.ResponseWriter) error
}

type DeleteLegalEntitiesUuid204Response struct {
}

func (response DeleteLegalEntitiesUuid204Response) VisitDeleteLegalEntitiesUuidResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteLegalEntitiesUuid401JSONResponse ErrorResponse

func (response DeleteLegalEntitiesUuid401JSONResponse) VisitDeleteLegalEntitiesUuidResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type PutLegalEntitiesUuidRequestObject struct {
	Uuid openapi_types.UUID `json:"uuid"`
	Body *PutLegalEntitiesUuidJSONRequestBody
}

type PutLegalEntitiesUuidResponseObject interface {
	VisitPutLegalEntitiesUuidResponse(w http.ResponseWriter) error
}

type PutLegalEntitiesUuid204Response struct {
}

func (response PutLegalEntitiesUuid204Response) VisitPutLegalEntitiesUuidResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PutLegalEntitiesUuid401JSONResponse ErrorResponse

func (response PutLegalEntitiesUuid401JSONResponse) VisitPutLegalEntitiesUuidResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Получить все юридические лица
	// (GET /legal-entities)
	GetLegalEntities(ctx context.Context, request GetLegalEntitiesRequestObject) (GetLegalEntitiesResponseObject, error)
	// Создать юридическое лицо
	// (POST /legal-entities)
	PostLegalEntities(ctx context.Context, request PostLegalEntitiesRequestObject) (PostLegalEntitiesResponseObject, error)
	// Удалить юридическое лицо
	// (DELETE /legal-entities/{uuid})
	DeleteLegalEntitiesUuid(ctx context.Context, request DeleteLegalEntitiesUuidRequestObject) (DeleteLegalEntitiesUuidResponseObject, error)
	// Обновить юридическое лицо
	// (PUT /legal-entities/{uuid})
	PutLegalEntitiesUuid(ctx context.Context, request PutLegalEntitiesUuidRequestObject) (PutLegalEntitiesUuidResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetLegalEntities operation middleware
func (sh *strictHandler) GetLegalEntities(ctx echo.Context) error {
	var request GetLegalEntitiesRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetLegalEntities(ctx.Request().Context(), request.(GetLegalEntitiesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetLegalEntities")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetLegalEntitiesResponseObject); ok {
		return validResponse.VisitGetLegalEntitiesResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostLegalEntities operation middleware
func (sh *strictHandler) PostLegalEntities(ctx echo.Context) error {
	var request PostLegalEntitiesRequestObject

	var body PostLegalEntitiesJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostLegalEntities(ctx.Request().Context(), request.(PostLegalEntitiesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostLegalEntities")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostLegalEntitiesResponseObject); ok {
		return validResponse.VisitPostLegalEntitiesResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteLegalEntitiesUuid operation middleware
func (sh *strictHandler) DeleteLegalEntitiesUuid(ctx echo.Context, uuid openapi_types.UUID) error {
	var request DeleteLegalEntitiesUuidRequestObject

	request.Uuid = uuid

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteLegalEntitiesUuid(ctx.Request().Context(), request.(DeleteLegalEntitiesUuidRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteLegalEntitiesUuid")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteLegalEntitiesUuidResponseObject); ok {
		return validResponse.VisitDeleteLegalEntitiesUuidResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PutLegalEntitiesUuid operation middleware
func (sh *strictHandler) PutLegalEntitiesUuid(ctx echo.Context, uuid openapi_types.UUID) error {
	var request PutLegalEntitiesUuidRequestObject

	request.Uuid = uuid

	var body PutLegalEntitiesUuidJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PutLegalEntitiesUuid(ctx.Request().Context(), request.(PutLegalEntitiesUuidRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PutLegalEntitiesUuid")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PutLegalEntitiesUuidResponseObject); ok {
		return validResponse.VisitPutLegalEntitiesUuidResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}