# openapi/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.5
  title: CRM API
  description: A main CRM Api
  contact:
    name: API Team
servers:
  - url: "localhost:8080"
    description: Dev server
...
paths:
...
  /legal-entities:
    get:
      operationId: getLegalEntities
      tags:
        - legal-entities
      summary: Получить все юридические лица
      security:
        - BearerAuth: []
      responses:
        200:
          description: Список юридических лиц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalEntityDTO"
        401:
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: postLegalEntities
      tags:
        - legal-entities
      summary: Создать юридическое лицо
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalEntityCreateRequest"
      responses:
        201:
          description: Юридическое лицо успешно создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLegalEntitiesResponse"
        401:
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /legal-entities/{uuid}:
    put:
      operationId: PutLegalEntitiesUuid
      tags:
        - legal-entities
      summary: Обновить юридическое лицо
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalEntityUpdateRequest"
      responses:
        204:
          description: Юридическое лицо успешно обновлено
        401:
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteLegalEntitiesUuid
      tags:
        - legal-entities
      summary: Удалить юридическое лицо
      security:
        - BearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Юридическое лицо успешно удалено
        401:
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"


#конец
...
components:
...
  schemas:
...
    LegalEntityDTO:
      type: object
      required:
        - uuid
        - name
        - created_at
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    LegalEntityCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    LegalEntityUpdateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    PostLegalEntitiesResponse:
      type: object
      required:
        - uuid
      properties:
        uuid:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    #конец DTO
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []